requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
google-genai
python-dotenv
psycopg2
